requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
dnspython==2.1.0
et-xmlfile==1.1.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
pymongo==4.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
splunk-handler==3.0.0
SQLAlchemy==1.4.29
tqdm==4.62.3
urllib3==1.26.8
